created reusable customisable app bar with search, currently implemented on home and view map screens
