+++ lib/screens/view_map_screen.dart
@@ -1,8 +1,8 @@
 // ignore_for_file: prefer_const_constructors, prefer_const_literals_to_create_immutables
 
 import 'package:flutter/material.dart';
+import 'package:tlydp/shared/search_appBar.dart';
 import '../shared/map.dart';
-import 'landing_screen.dart';
 
 class ViewMap extends StatelessWidget {
   const ViewMap({Key? key}) : super(key: key);
@@ -10,23 +10,10 @@ class ViewMap extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        appBar: AppBar(
-            backgroundColor: Color.fromARGB(255, 223, 105, 8),
-            leading: IconButton(
-              icon: const Icon(Icons.arrow_back),
-              onPressed: () {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(
-                      builder: (context) => const LandingScreen()),
-                );
-              },
-            ),
-            title: Text('The Little Yellow Duck Project',
-                style: TextStyle(
-                    fontSize: 20,
-                    fontWeight: FontWeight.w800,
-                    color: Colors.white))),
+        appBar: SearchAppBar(
+          title: 'The Little Yellow Duck Project',
+          showMenu: false,
+        ),
         body: Center(child: Map()));
   }
 }
